feat: implement user authentication flow with login/register pages and protected routes
