+++ nextjs-frontend/app/loading.tsx
@@ -1,3 +1,4 @@
+// Intentionally minimal - NextTopLoader in layout.tsx handles all loading UI
 export default function Loading() {
-  return null
+  return null;
 }
