+++ nextjs-frontend/app/login/page.tsx
@@ -1,6 +1,7 @@
 "use client";
 
 import Link from "next/link";
+import { usePathname } from "next/navigation";
 import {
   Card,
   CardContent,
@@ -10,80 +11,112 @@ import {
 } from "@/components/ui/card";
 import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
+import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
+import { Button } from "@/components/ui/button";
+import { Mail, Lock, FileText } from "lucide-react";
 
 import { login } from "@/components/actions/login-action";
 import { useActionState } from "react";
-import { SubmitButton } from "@/components/ui/submitButton";
 import { FieldError, FormError } from "@/components/ui/FormError";
 
 export default function Page() {
+  const pathname = usePathname();
   const [state, dispatch] = useActionState(login, undefined);
+  
   return (
-    <div className="flex h-screen w-full items-center justify-center bg-gray-50 dark:bg-gray-900 px-4">
-      <form action={dispatch}>
-        <Card className="w-full max-w-sm rounded-lg shadow-lg border border-gray-300 dark:border-gray-700 bg-white dark:bg-gray-800">
-          <CardHeader className="text-center">
-            <CardTitle className="text-2xl font-semibold text-gray-800 dark:text-white">
-              Login
-            </CardTitle>
-            <CardDescription className="text-sm text-gray-600 dark:text-gray-400">
-              Enter your email below to log in to your account.
+    <div className="min-h-screen bg-background flex items-center justify-center p-4">
+      <div className="w-full max-w-md">
+        <div className="text-center mb-8">
+          <div className="flex items-center justify-center gap-2 mb-4">
+            <div className="p-2 bg-gradient-to-r from-orange-500 to-yellow-500 rounded-xl">
+              <FileText className="h-8 w-8 text-white" />
+            </div>
+            <h1 className="text-3xl font-bold bg-gradient-to-r from-orange-600 to-yellow-600 bg-clip-text text-transparent">
+              ColPali
+            </h1>
+          </div>
+          <p className="text-gray-600 text-lg">Document Search Platform</p>
+        </div>
+        
+        <Card className="shadow-xl border border-border bg-card">
+          <CardHeader className="text-center pb-4">
+            <CardTitle className="text-2xl font-bold text-card-foreground">Welcome</CardTitle>
+            <CardDescription className="text-muted-foreground">
+              Sign in to your account to get started
             </CardDescription>
           </CardHeader>
-          <CardContent className="grid gap-6 p-6">
-            <div className="grid gap-3">
-              <Label
-                htmlFor="username"
-                className="text-gray-700 dark:text-gray-300"
-              >
-                Username
-              </Label>
-              <Input
-                id="username"
-                name="username"
-                type="email"
-                placeholder="m@example.com"
-                required
-                className="border-gray-300 dark:border-gray-600"
-              />
-              <FieldError state={state} field="username" />
-            </div>
-            <div className="grid gap-3">
-              <Label
-                htmlFor="password"
-                className="text-gray-700 dark:text-gray-300"
-              >
-                Password
-              </Label>
-              <Input
-                id="password"
-                name="password"
-                type="password"
-                required
-                className="border-gray-300 dark:border-gray-600"
-              />
-              <FieldError state={state} field="password" />
-              <Link
-                href="/password-recovery"
-                className="ml-auto inline-block text-sm text-blue-500 hover:text-blue-600 dark:text-blue-400 dark:hover:text-blue-500"
-              >
-                Forgot your password?
-              </Link>
-            </div>
-            <SubmitButton text="Sign In" />
-            <FormError state={state} />
-            <div className="mt-4 text-center text-sm text-gray-600 dark:text-gray-400">
-              Don&apos;t have an account?{" "}
-              <Link
-                href="/register"
-                className="text-blue-500 hover:text-blue-600 dark:text-blue-400 dark:hover:text-blue-500"
-              >
-                Sign up
-              </Link>
+          <CardContent>
+            <div className="flex flex-col space-y-6">
+              <div className="flex w-full rounded-lg overflow-hidden border border-border bg-muted">
+                <Link 
+                  href="/login" 
+                  className={`flex-1 py-2 text-center font-medium ${pathname === '/login' ? 'bg-gradient-to-r from-orange-500 to-yellow-500 text-white' : 'text-muted-foreground'}`}
+                >
+                  Login
+                </Link>
+                <Link 
+                  href="/register" 
+                  className={`flex-1 py-2 text-center font-medium ${pathname === '/register' ? 'bg-gradient-to-r from-orange-500 to-yellow-500 text-white' : 'text-muted-foreground'}`}
+                >
+                  Register
+                </Link>
+              </div>
+              
+              <div>
+                <form action={dispatch} className="space-y-4">
+                  <div className="space-y-2">
+                    <Label htmlFor="username" className="text-card-foreground font-medium">
+                      Email
+                    </Label>
+                    <div className="relative">
+                      <Mail className="absolute left-3 top-3 h-4 w-4 text-gray-400" />
+                      <Input
+                        id="username"
+                        name="username"
+                        type="email"
+                        placeholder="Enter your email"
+                        className="pl-10 border-border focus:border-orange-400 focus:ring-orange-400 bg-background text-foreground"
+                        required
+                      />
+                    </div>
+                    <FieldError state={state} field="username" />
+                  </div>
+                  <div className="space-y-2">
+                    <Label htmlFor="password" className="text-card-foreground font-medium">
+                      Password
+                    </Label>
+                    <div className="relative">
+                      <Lock className="absolute left-3 top-3 h-4 w-4 text-gray-400" />
+                      <Input
+                        id="password"
+                        name="password"
+                        type="password"
+                        placeholder="Enter your password"
+                        className="pl-10 border-border focus:border-orange-400 focus:ring-orange-400 bg-background text-foreground"
+                        required
+                      />
+                    </div>
+                    <FieldError state={state} field="password" />
+                    <Link
+                      href="/password-recovery"
+                      className="block text-right text-sm text-orange-500 hover:text-orange-600 mt-1"
+                    >
+                      Forgot your password?
+                    </Link>
+                  </div>
+                  <Button
+                    type="submit"
+                    className="w-full bg-gradient-to-r from-orange-500 to-yellow-500 hover:from-orange-600 hover:to-yellow-600 text-white"
+                  >
+                    Sign In
+                  </Button>
+                  <FormError state={state} />
+                </form>
+              </div>
             </div>
           </CardContent>
         </Card>
-      </form>
+      </div>
     </div>
   );
 }
