+++ nextjs-frontend/app/page.tsx
@@ -1,233 +1,31 @@
-"use client"
+"use client";
 
-import { useState } from "react"
-import { AppSidebar } from "@/components/app-sidebar"
-import { DocumentUpload } from "@/components/document-upload"
-import { DocumentSearch } from "@/components/document-search"
-import { CollectionInfo } from "@/components/collection-info"
-import { UserProfile } from "@/components/user-profile"
-import { SidebarProvider, SidebarInset, SidebarTrigger } from "@/components/ui/sidebar"
-import { Separator } from "@/components/ui/separator"
-import { Button } from "@/components/ui/button"
-import { Input } from "@/components/ui/input"
-import { Label } from "@/components/ui/label"
-import { Card, CardContent, CardDescription, CardHeader, CardTitle } from "@/components/ui/card"
-import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs"
-import { FileText, Mail, Lock, User } from "lucide-react"
-import { useRouter } from "next/navigation"
+import { useEffect } from 'react';
+import { useRouter } from 'next/navigation';
 
-export default function Dashboard() {
-  const [activeTab, setActiveTab] = useState("search")
-  const [isAuthenticated, setIsAuthenticated] = useState(false)
-  const [isLoading, setIsLoading] = useState(false)
-  const router = useRouter()
+// This is a simple page that redirects to the appropriate page based on auth status
+export default function Home() {
+  const router = useRouter();
 
-  const handleLogin = async (e: React.FormEvent) => {
-    e.preventDefault()
-    setIsLoading(true)
-    // TODO: Replace with actual authentication logic
-    try {
-      // Simulate API call
-      await new Promise(resolve => setTimeout(resolve, 1000))
-      setIsAuthenticated(true)
-      router.push('/')
-    } catch (error) {
-      console.error('Login failed:', error)
-    } finally {
-      setIsLoading(false)
+  useEffect(() => {
+    // Check if we have an access token in client-side cookies
+    const accessToken = document.cookie
+      .split('; ')
+      .find((row) => row.startsWith('accessToken='))
+      ?.split('=')[1];
+    
+    // If there's a token, redirect to dashboard, otherwise to login
+    if (accessToken) {
+      router.push('/dashboard');
+    } else {
+      router.push('/login');
     }
-  }
+  }, [router]);
 
-  const handleRegister = async (e: React.FormEvent) => {
-    e.preventDefault()
-    setIsLoading(true)
-    // TODO: Replace with actual registration logic
-    try {
-      // Simulate API call
-      await new Promise(resolve => setTimeout(resolve, 1000))
-      setIsAuthenticated(true)
-      router.push('/')
-    } catch (error) {
-      console.error('Registration failed:', error)
-    } finally {
-      setIsLoading(false)
-    }
-  }
-
-  if (!isAuthenticated) {
-    return (
-      <div className="min-h-screen bg-background flex items-center justify-center p-4">
-        <div className="w-full max-w-md">
-          <div className="text-center mb-8">
-            <div className="flex items-center justify-center gap-2 mb-4">
-              <div className="p-2 bg-gradient-to-r from-orange-500 to-yellow-500 rounded-xl">
-                <FileText className="h-8 w-8 text-white" />
-              </div>
-              <h1 className="text-3xl font-bold bg-gradient-to-r from-orange-600 to-yellow-600 bg-clip-text text-transparent">
-                ColPali
-              </h1>
-            </div>
-            <p className="text-gray-600 text-lg">Document Search Platform</p>
-          </div>
-
-          <Card className="shadow-xl border border-border bg-card">
-            <CardHeader className="text-center pb-4">
-              <CardTitle className="text-2xl font-bold text-card-foreground">Welcome</CardTitle>
-              <CardDescription className="text-muted-foreground">
-                Sign in to your account or create a new one to get started
-              </CardDescription>
-            </CardHeader>
-            <CardContent>
-              <Tabs defaultValue="login" className="w-full">
-                <TabsList className="grid w-full grid-cols-2 mb-6 bg-muted">
-                  <TabsTrigger
-                    value="login"
-                    className="data-[state=active]:bg-gradient-to-r data-[state=active]:from-orange-500 data-[state=active]:to-yellow-500 data-[state=active]:text-white"
-                  >
-                    Login
-                  </TabsTrigger>
-                  <TabsTrigger
-                    value="register"
-                    className="data-[state=active]:bg-gradient-to-r data-[state=active]:from-orange-500 data-[state=active]:to-yellow-500 data-[state=active]:text-white"
-                  >
-                    Register
-                  </TabsTrigger>
-                </TabsList>
-
-                <TabsContent value="login">
-                  <form onSubmit={handleLogin} className="space-y-4">
-                    <div className="space-y-2">
-                      <Label htmlFor="email" className="text-card-foreground font-medium">
-                        Email
-                      </Label>
-                      <div className="relative">
-                        <Mail className="absolute left-3 top-3 h-4 w-4 text-gray-400" />
-                        <Input
-                          id="email"
-                          type="email"
-                          placeholder="Enter your email"
-                          className="pl-10 border-border focus:border-orange-400 focus:ring-orange-400 bg-background text-foreground"
-                          required
-                        />
-                      </div>
-                    </div>
-                    <div className="space-y-2">
-                      <Label htmlFor="password" className="text-card-foreground font-medium">
-                        Password
-                      </Label>
-                      <div className="relative">
-                        <Lock className="absolute left-3 top-3 h-4 w-4 text-gray-400" />
-                        <Input
-                          id="password"
-                          type="password"
-                          placeholder="Enter your password"
-                          className="pl-10 border-border focus:border-orange-400 focus:ring-orange-400 bg-background text-foreground"
-                          required
-                        />
-                      </div>
-                    </div>
-                    <Button
-                      type="submit"
-                      className="w-full bg-gradient-to-r from-orange-500 to-yellow-500 hover:from-orange-600 hover:to-yellow-600 text-white"
-                      disabled={isLoading}
-                    >
-                      {isLoading ? 'Signing in...' : 'Sign In'}
-                    </Button>
-                  </form>
-                </TabsContent>
-
-                <TabsContent value="register">
-                  <form onSubmit={handleRegister} className="space-y-4">
-                    <div className="space-y-2">
-                      <Label htmlFor="name" className="text-card-foreground font-medium">
-                        Full Name
-                      </Label>
-                      <div className="relative">
-                        <User className="absolute left-3 top-3 h-4 w-4 text-gray-400" />
-                        <Input
-                          id="name"
-                          type="text"
-                          placeholder="Enter your full name"
-                          className="pl-10 border-border focus:border-orange-400 focus:ring-orange-400 bg-background text-foreground"
-                          required
-                        />
-                      </div>
-                    </div>
-                    <div className="space-y-2">
-                      <Label htmlFor="reg-email" className="text-card-foreground font-medium">
-                        Email
-                      </Label>
-                      <div className="relative">
-                        <Mail className="absolute left-3 top-3 h-4 w-4 text-gray-400" />
-                        <Input
-                          id="reg-email"
-                          type="email"
-                          placeholder="Enter your email"
-                          className="pl-10 border-border focus:border-orange-400 focus:ring-orange-400 bg-background text-foreground"
-                          required
-                        />
-                      </div>
-                    </div>
-                    <div className="space-y-2">
-                      <Label htmlFor="reg-password" className="text-card-foreground font-medium">
-                        Password
-                      </Label>
-                      <div className="relative">
-                        <Lock className="absolute left-3 top-3 h-4 w-4 text-gray-400" />
-                        <Input
-                          id="reg-password"
-                          type="password"
-                          placeholder="Create a password"
-                          className="pl-10 border-border focus:border-orange-400 focus:ring-orange-400 bg-background text-foreground"
-                          required
-                        />
-                      </div>
-                    </div>
-                    <Button
-                      type="submit"
-                      className="w-full bg-gradient-to-r from-orange-500 to-yellow-500 hover:from-orange-600 hover:to-yellow-600 text-white"
-                      disabled={isLoading}
-                    >
-                      {isLoading ? 'Creating account...' : 'Create Account'}
-                    </Button>
-                  </form>
-                </TabsContent>
-              </Tabs>
-            </CardContent>
-          </Card>
-        </div>
-      </div>
-    )
-  }
-
-  const renderContent = () => {
-    switch (activeTab) {
-      case "search":
-        return <DocumentSearch />
-      case "upload":
-        return <DocumentUpload />
-      case "collection":
-        return <CollectionInfo />
-      case "profile":
-        return <UserProfile />
-      default:
-        return <DocumentSearch />
-    }
-  }
-
+  // Return a minimal UI while redirecting - NextTopLoader will show a progress bar
   return (
-    <SidebarProvider>
-      <AppSidebar activeTab={activeTab} onTabChange={setActiveTab} />
-      <SidebarInset>
-        <header className="flex h-16 shrink-0 items-center gap-2 border-b px-4">
-          <SidebarTrigger className="-ml-1" />
-          <Separator orientation="vertical" className="mr-2 h-4" />
-          <div className="flex items-center gap-2">
-            <h1 className="text-lg font-semibold">ColPali Dashboard</h1>
-          </div>
-        </header>
-        <main className="flex-1 overflow-hidden">{renderContent()}</main>
-      </SidebarInset>
-    </SidebarProvider>
-  )
+    <div className="min-h-screen bg-background">
+      {/* Intentionally empty, NextTopLoader will provide visual feedback */}
+    </div>
+  );
 }
