+++ nextjs-frontend/middleware.ts
@@ -24,5 +24,5 @@ export async function middleware(request: NextRequest) {
 }
 
 export const config = {
-  matcher: ["/:path*"],
+  matcher: ["/dashboard/:path*"],
 };
