+++ nextjs-frontend/package.json
@@ -60,6 +60,7 @@
     "lucide-react": "^0.452.0",
     "next": "15.2.4",
     "next-themes": "^0.4.6",
+    "nextjs-toploader": "^3.8.16",
     "react": "19.0.0",
     "react-day-picker": "^9.8.1",
     "react-dom": "19.0.0",
